pyproject.toml
[tool.poetry]
name = "image-coloring"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "3.8.10"
opencv-python = "^4.5.3"
numpy = "^1.21.2"
Flask = "^2.0.2"
Pillow = "^8.3.2"
requests = "^2.26.0"
Flask-Cors = "^3.0.10"



[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
